SplitWorkload/backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.1
pandas==2.2.0
openpyxl==3.1.2
python-multipart==0.0.9
pydantic==2.6.1
numpy==1.26.3
httpx==0.26.0
cachetools==5.3.2
python-dotenv==1.0.0
jinja2==3.1.4
